+++ poc/classes/AuxSTVarSpecList.py
@@ -1,6 +1,5 @@
 from poc.classes.AuxEvaluation import EvaluateParams
 from poc.classes.AuxInbuiltTypes import InbuiltUndefined
-from poc.classes.AuxSTVarDec import AuxSTVarDec
 from poc.classes.AuxST import AuxSTOutline
 from poc.classes.AuxSTConstants import AuxSTConstants
 
@@ -17,12 +16,19 @@ class AuxSTVarSpecList(AuxSTOutline):
 
     def evaluate(self, sem):
         for child in self.children:
-            if isinstance(child, AuxSTVarDec):
-                # initialize the declared type of AuxSTVarDec nodes
-                child.set_declared_type(InbuiltUndefined(child))
             expected_type = self._determine_exptected_type(child)
             EvaluateParams.evaluate_recursion(sem, child, expected_type)
         sem.eval_stack[-1].value = InbuiltUndefined(self.parent)
 
     def _determine_exptected_type(self, child):
-        return InbuiltUndefined(child)
+        if child.outline == AuxSTConstants.statement_assign:
+            # the expected type of the assignment is the type of the variable to which we assign the value
+            return child.children[0].get_declared_type()
+        elif child.outline == AuxSTConstants.statement_return:
+            return child.get_declared_type()
+        elif child.outline == AuxSTConstants.var_decl:
+            # the expected type of var declarations is undefined, which is also initialize
+            child.set_declared_type(InbuiltUndefined(child))
+            return child.get_declared_type()
+        else:
+            raise NotImplementedError(child)