+++ poc/tests/errors_signatures/test_AllFplSignatureRelatedErrorTests.py
@@ -18,6 +18,8 @@ class AllFplSignatureRelatedErrorTests(UtilTestCase):
         ("test_FplWrongArguments_03b.fpl", "SE0220"),
         ("test_FplWrongArguments_04.fpl", "SE0220"),
         ("test_FplWrongArguments_05.fpl", "SE0220"),
+        ("test_FplInvalidUseReturnStmt_01.fpl", "SE0290"),
+        ("test_FplInvalidUseReturnStmt_02.fpl", "SE0290"),
     ])
     def test_errors(self, use_case, diagnose_id):
         super().semantical_analysis_detects_fpl_error(self.folder + "/" + use_case, diagnose_id)
@@ -38,6 +40,7 @@ class AllFplSignatureRelatedErrorTests(UtilTestCase):
         ("test_FplWrongArguments_ok_05d.fpl", "SE0220"),
         ("test_FplWrongArguments_ok_06.fpl", "SE0220"),
         ("test_FplWrongArguments_ok_07.fpl", "SE0220"),
+        ("test_FplInvalidUseReturnStmt_ok_01.fpl", "SE0290"),
     ])
     def test_no_errors(self, use_case, diagnose_id):
         super().semantical_analysis_detects_no_fpl_error(self.folder + "/" + use_case, diagnose_id)